More progress

+++ source/engine/classes/entitymanager.lua
@@ -4,7 +4,6 @@ local EntityManager = class("EntityManager")
 function EntityManager:initialize()
 	
 	self._entities = {}
-	--print(table.toString(self._entities, "entities", true))
 	self._drawlist = {}
 	self._update_drawlist = true
 	
@@ -44,19 +43,22 @@ function EntityManager:createEntity( class, ...)
 	
 end
 
-function EntityManager:removeEntity( ent )
-	
-	table.removeByValue(self._entities, ent)
-	ent:onRemove()
-	
-	self._update_drawlist = true
-	
-end
-
 function EntityManager:update( dt )
 	
-	for k, v in ipairs( self._entities ) do
-		v:update( dt )
+	local ent
+	-- traverse table in reverse order so we can safely remove entities while traversing
+	for i = #self._entities, 1, -1 do
+		
+		ent = self._entities[i]
+		ent:update( dt )
+		
+		if (ent._removeflag) then
+			table.remove(self._entities, i)
+			ent:onRemove()
+			
+			self._update_drawlist = true
+		end		
+		
 	end
 	
 end
